Il peut arriver que le framerate soit un float...
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@41071 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ spipmotion/base/spipmotion.php
@@ -38,7 +38,7 @@ function spipmotion_declarer_tables_principales($tables_principales){
 	$tables_principales['spip_documents']['field']['duree'] = "VARCHAR(255) DEFAULT '' NOT NULL";
 	$tables_principales['spip_documents']['field']['hasvideo'] = "VARCHAR(3) DEFAULT '' NOT NULL";
 	$tables_principales['spip_documents']['field']['framecount'] = "INTEGER";
-	$tables_principales['spip_documents']['field']['framerate'] = "INTEGER";
+	$tables_principales['spip_documents']['field']['framerate'] = "FLOAT";
 	$tables_principales['spip_documents']['field']['pixelformat'] = "VARCHAR(255) DEFAULT '' NOT NULL";
 	$tables_principales['spip_documents']['field']['bitrate'] = "INTEGER";
 	$tables_principales['spip_documents']['field']['videobitrate'] = "INTEGER";
